Widen plane samples from 8-bit to 16-bit storage
First step towards 10/12-bit stream support: Av1Plane and every
sample-carrying path (intra prediction, directional/filter-intra
edges, CfL, motion compensation, warp, OBMC and compound blends,
deblocking, CDEF, loop restoration) now operate on 16-bit samples.
Weight, mask and filter-tap tables keep their 8-bit types; 8-bit
decode behaviour is unchanged and clamps remain at 255 until the
kernels are parametrised by bit depth.

The full regression suite (508 tests, byte-exact against dav1d)
passes unchanged on the widened representation.

+++ src/SixLabors.ImageSharp.Video/Av1FrameConverter.cs
@@ -65,7 +65,7 @@ internal static class Av1FrameConverter
     }
 
     // BT.601 limited-range YUV to RGB conversion (specification's default matrix for 8-bit content).
-    private static void YuvToRgb(byte yy, byte uu, byte vv, ref Rgba32 rgba)
+    private static void YuvToRgb(ushort yy, ushort uu, ushort vv, ref Rgba32 rgba)
     {
         float y = 1.164f * (yy - 16);
         float u = uu - 128;
