+++ src/SixLabors.ImageSharp.Video/Bitstream/Av1Plane.cs
@@ -4,14 +4,14 @@
 namespace SixLabors.ImageSharp.Formats.Av1.Bitstream;
 
 /// <summary>
-/// A single reconstructed image plane (8-bit samples) backed by a dense row-major buffer. The allocated
+/// A single reconstructed image plane (16-bit sample storage, holding 8-, 10- or 12-bit values) backed by a dense row-major buffer. The allocated
 /// area covers the frame's whole 4x4 block grid so transform blocks reconstruct fully even when they
 /// overhang the visible frame (the reference decoder pads its planes the same way); the crop dimensions
 /// give the visible size used for output and for motion-compensation edge replication.
 /// </summary>
 internal sealed class Av1Plane
 {
-    private readonly byte[] samples;
+    private readonly ushort[] samples;
 
     public Av1Plane(int width, int height)
         : this(width, height, width, height)
@@ -24,7 +24,7 @@ internal sealed class Av1Plane
         this.Height = height;
         this.CropWidth = cropWidth;
         this.CropHeight = cropHeight;
-        this.samples = new byte[width * height];
+        this.samples = new ushort[width * height];
     }
 
     /// <summary>Gets the allocated plane width in samples (the row stride).</summary>
@@ -40,12 +40,12 @@ internal sealed class Av1Plane
     public int CropHeight { get; }
 
     /// <summary>Gets the backing sample buffer in row-major order.</summary>
-    public byte[] Samples => this.samples;
+    public ushort[] Samples => this.samples;
 
     /// <summary>Gets or sets the sample at the given coordinate.</summary>
     /// <param name="x">The column.</param>
     /// <param name="y">The row.</param>
-    public byte this[int x, int y]
+    public ushort this[int x, int y]
     {
         get => this.samples[(y * this.Width) + x];
         set => this.samples[(y * this.Width) + x] = value;
